qpnp: Correct the variable declaration for QPNP_POFF_REASON1 reg read
When reading QPNP_POFF_REASON1 to provide the last power off reason
through /proc/sys/kernel/poweroff_reason to userspace, we found that
it may provide the wrong reason sometimes, such as cut off the power
and then power on.
The value read back from the PMIC register is 8 bits. If using unsigned
int buf[2] to read back 2 byets start from QPNP_POFF_REASON1 reg and then
buf[0] | (buf[1] << 8), it will overwirte the second byte value.

So Correct the variable declaration to u8 for QPNP_POFF_REASON1 reg read.

Resolves: LXSWIREF-2016
Relates: QTI9X07-5393

+++ drivers/input/misc/qpnp-power-on.c
@@ -2149,7 +2149,7 @@ static int qpnp_pon_read_hardware_info(struct qpnp_pon *pon, bool sys_reset)
 {
 	struct device *dev = pon->dev;
 	unsigned int reg = 0;
-	unsigned int buf[2];
+	u8 buf[2];
 	int reason_index_offset = 0;
 	unsigned int pon_sts = 0;
 	u16 poff_sts = 0;
